requirements.txt
# Google APIs for Gmail
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0

# Anthropic Claude API
anthropic>=0.44.0
httpx>=0.27.0

# HTML parsing
beautifulsoup4==4.12.2
lxml==5.1.0

# Templating
jinja2==3.1.3

# Environment variables
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# YAML config (optional)
pyyaml==6.0.1
